+++ examples/workflow-server/src/app.ts
@@ -13,8 +13,10 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
+import { configureELKLayoutModule } from '@eclipse-glsp/layout-elk';
 import { createAppModule, createSocketCliParser, LoggerFactory, resolveAndCatch, SocketServerLauncher } from '@eclipse-glsp/server-node';
 import { Container } from 'inversify';
+import { WorkflowLayoutConfigurator } from './layout/workflow-layout-configurator';
 import { WorkflowDiagramModule, WorkflowServerModule } from './workflow-diagram-module';
 
 export function launch(argv?: string[]): void {
@@ -24,7 +26,8 @@ export function launch(argv?: string[]): void {
 
     const logger = appContainer.get<LoggerFactory>(LoggerFactory)('WorkflowServerApp');
     const launcher = appContainer.resolve(SocketServerLauncher);
-    const serverModule = new WorkflowServerModule().configureDiagramModule(new WorkflowDiagramModule());
+    const elkLayoutModule = configureELKLayoutModule({ algorithms: ['layered'], layoutConfigurator: WorkflowLayoutConfigurator });
+    const serverModule = new WorkflowServerModule().configureDiagramModule(new WorkflowDiagramModule(), elkLayoutModule);
 
     const errorHandler = (error: any): void => logger.error('Error in workflow server launcher:', error);
     launcher.configure(serverModule);